Rename bot to JerryBot
