+++ src/main/java/jerryBot/JerryBot.java
@@ -1,4 +1,4 @@
-package botchat;
+package jerryBot;
 
 import java.util.ArrayList;
 import java.util.Scanner;
@@ -6,13 +6,6 @@ import java.util.Scanner;
 import exception.IncompleteCommandException;
 import exception.InvalidCommandException;
 import exception.InvalidTaskNumberException;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.layout.VBox;
 import parser.Parser;
 import storage.Storage;
 import task.Deadline;
@@ -26,14 +19,17 @@ import ui.Ui;
  * This class is the chatbot.
  *
  */
-public class BotChat {
+public class JerryBot {
     private static boolean isTerminate = false;
-    private static final String FILEPATH = "./././data/botchat.txt";
+    private static final String FILEPATH = "./././data/jerrybot.txt";
     private static Storage storage;
     private static TaskList taskArrayList;
     private static Parser parser;
 
-    public BotChat() {
+    /**
+     * Constructor to initialise the chatbot class.
+     */
+    public JerryBot() {
         parser = new Parser();
         storage = new Storage(FILEPATH);
         taskArrayList = new TaskList(storage.readDataStore());
@@ -213,17 +209,6 @@ public class BotChat {
         } catch (NumberFormatException e) {
             throw new InvalidTaskNumberException(requestedDeletion);
         }
-    }
-
-    private Label getDialogLabel(String text) {
-        Label textToAdd = new Label(text);
-        textToAdd.setWrapText(true);
-
-        return textToAdd;
-    }
-
-    public String getResponse(String input) {
-        return response(input);
     }
 
     public static void main(String[] args) {
