+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm botchat.BotChat.
+Hello! I'm jerryBot.BotChat.
  What can I do for you?
 Invalid command erm. Please try again with todo, deadline event or list.
 Todo command incomplete. It should be in the form of todo description.
